requirements.txt
feedparser
google-generativeai
deep-translator>=1.11,<2.0
# Pin to a version that exposes the `Groq` client class used in generate_news.py
groq>=0.4.2,<1.0
PyYAML
Jinja2
